draft: first draft of sdk implementation
