+++ core/opaevaluator.go
@@ -15,13 +15,11 @@
 package core
 
 import (
-	"bytes"
 	"context"
 	"encoding/json"
 	"fmt"
 	"io"
 	"net/http"
-	"net/url"
 	"os"
 	"path/filepath"
 	"strings"
@@ -378,27 +376,6 @@ func (evaluator *OPAEvaluator) PolicyEvaluation(logger *logrus.Entry, permission
 	return dataFromEvaluation, nil, nil
 }
 
-type RegoInputOptions struct {
-	EnableResourcePermissionsMapOptimization bool
-}
-
-func CreateRegoQueryInput(
-	logger *logrus.Entry,
-	input Input,
-	options RegoInputOptions,
-) ([]byte, error) {
-	opaInputCreationTime := time.Now()
-
-	input.buildOptimizedResourcePermissionsMap(logger, options.EnableResourcePermissionsMapOptimization)
-
-	inputBytes, err := json.Marshal(input)
-	if err != nil {
-		return nil, fmt.Errorf("failed input JSON encode: %v", err)
-	}
-	logger.Tracef("OPA input rego creation in: %+v", time.Since(opaInputCreationTime))
-	return inputBytes, nil
-}
-
 func buildRolesMap(roles []types.Role) map[string][]string {
 	var rolesMap = make(map[string][]string, 0)
 	for _, role := range roles {
@@ -443,64 +420,6 @@ func GetOPAModuleConfig(requestContext context.Context) (*OPAModuleConfig, error
 	return permission, nil
 }
 
-type Input struct {
-	Request    InputRequest  `json:"request"`
-	Response   InputResponse `json:"response"`
-	ClientType string        `json:"clientType,omitempty"`
-	User       InputUser     `json:"user"`
-}
-
-func (input *Input) buildOptimizedResourcePermissionsMap(logger *logrus.Entry, enableResourcePermissionsMapOptimization bool) {
-	if !enableResourcePermissionsMapOptimization {
-		return
-	}
-	logger.Info("preparing optimized resourcePermissionMap for OPA evaluator")
-	opaPermissionsMapTime := time.Now()
-
-	user := input.User
-	permissionsOnResourceMap := make(PermissionsOnResourceMap, 0)
-	rolesMap := buildRolesMap(user.Roles)
-	for _, binding := range user.Bindings {
-		for _, role := range binding.Roles {
-			rolePermissions, ok := rolesMap[role]
-			if !ok {
-				continue
-			}
-			for _, permission := range rolePermissions {
-				key := buildPermissionOnResourceKey(permission, binding.Resource.ResourceType, binding.Resource.ResourceID)
-				permissionsOnResourceMap[key] = true
-			}
-		}
-		for _, permission := range binding.Permissions {
-			key := buildPermissionOnResourceKey(permission, binding.Resource.ResourceType, binding.Resource.ResourceID)
-			permissionsOnResourceMap[key] = true
-		}
-	}
-	input.User.ResourcePermissionsMap = permissionsOnResourceMap
-	logger.WithField("resourcePermissionMapCreationTime", fmt.Sprintf("%+v", time.Since(opaPermissionsMapTime))).Tracef("resource permission map creation")
-}
-
-type InputRequest struct {
-	Body       interface{}       `json:"body,omitempty"`
-	Headers    http.Header       `json:"headers,omitempty"`
-	Query      url.Values        `json:"query,omitempty"`
-	PathParams map[string]string `json:"pathParams,omitempty"`
-	Method     string            `json:"method"`
-	Path       string            `json:"path"`
-}
-
-type InputResponse struct {
-	Body interface{} `json:"body,omitempty"`
-}
-
-type InputUser struct {
-	Properties             map[string]interface{}   `json:"properties,omitempty"`
-	Groups                 []string                 `json:"groups,omitempty"`
-	Bindings               []types.Binding          `json:"bindings,omitempty"`
-	Roles                  []types.Role             `json:"roles,omitempty"`
-	ResourcePermissionsMap PermissionsOnResourceMap `json:"resourcePermissionsMap,omitempty"`
-}
-
 type PermissionOnResourceKey string
 
 type PermissionsOnResourceMap map[PermissionOnResourceKey]bool
@@ -537,50 +456,5 @@ func LoadRegoModule(rootDirectory string) (*OPAModuleConfig, error) {
 	return &OPAModuleConfig{
 		Name:    filepath.Base(regoModulePath),
 		Content: string(fileContent),
-	}, nil
-}
-
-func InputFromRequest(
-	req *http.Request,
-	user types.User,
-	clientTypeHeaderKey string,
-	pathParams map[string]string,
-	responseBody any,
-) (Input, error) {
-	shouldParseJSONBody := utils.HasApplicationJSONContentType(req.Header) &&
-		req.ContentLength > 0 &&
-		(req.Method == http.MethodPatch || req.Method == http.MethodPost || req.Method == http.MethodPut || req.Method == http.MethodDelete)
-
-	var requestBody any
-	if shouldParseJSONBody {
-		bodyBytes, err := io.ReadAll(req.Body)
-		if err != nil {
-			return Input{}, fmt.Errorf("failed request body parse: %s", err.Error())
-		}
-		if err := json.Unmarshal(bodyBytes, &requestBody); err != nil {
-			return Input{}, fmt.Errorf("failed request body deserialization: %s", err.Error())
-		}
-		req.Body = io.NopCloser(bytes.NewBuffer(bodyBytes))
-	}
-
-	return Input{
-		ClientType: req.Header.Get(clientTypeHeaderKey),
-		Request: InputRequest{
-			Method:     req.Method,
-			Path:       req.URL.Path,
-			Headers:    req.Header,
-			Query:      req.URL.Query(),
-			PathParams: pathParams,
-			Body:       requestBody,
-		},
-		Response: InputResponse{
-			Body: responseBody,
-		},
-		User: InputUser{
-			Properties: user.Properties,
-			Groups:     user.UserGroups,
-			Bindings:   user.UserBindings,
-			Roles:      user.UserRoles,
-		},
 	}, nil
 }
