+++ service/handler.go
@@ -204,7 +204,7 @@ func ReverseProxy(
 			req.URL.Scheme = URL_SCHEME
 			if _, ok := req.Header["User-Agent"]; !ok {
 				// explicitly disable User-Agent so it's not set to default value
-				req.Header.Set("User-Agent", "")
+				req.Header.Del("User-Agent")
 			}
 		},
 	}
