+++ service/statusroutes_test.go
@@ -27,6 +27,7 @@ import (
 	"github.com/rond-authz/rond/internal/config"
 	"github.com/rond-authz/rond/internal/mongoclient"
 	"github.com/rond-authz/rond/openapi"
+	"github.com/rond-authz/rond/sdk"
 
 	"github.com/gorilla/mux"
 	"github.com/sirupsen/logrus"
@@ -112,7 +113,7 @@ test_policy { true }
 	}
 
 	var mongoClient *mongoclient.MongoClient
-	evaluatorsMap, err := core.SetupEvaluators(ctx, mongoClient, oas, opa, nil)
+	sdk, err := sdk.New(ctx, mongoClient, oas, opa, nil, nil)
 	require.NoError(t, err, "unexpected error")
 
 	t.Run("non standalone", func(t *testing.T) {
@@ -121,7 +122,7 @@ test_policy { true }
 			TargetServiceHost:    "my-service:4444",
 			PathPrefixStandalone: "/my-prefix",
 		}
-		router, err := SetupRouter(log, env, opa, oas, evaluatorsMap, mongoClient)
+		router, err := SetupRouter(log, env, opa, oas, sdk, mongoClient)
 		require.NoError(t, err, "unexpected error")
 
 		t.Run("/-/rbac-ready", func(t *testing.T) {
@@ -154,7 +155,7 @@ test_policy { true }
 			PathPrefixStandalone: "/my-prefix",
 			ServiceVersion:       "latest",
 		}
-		router, err := SetupRouter(log, env, opa, oas, evaluatorsMap, mongoClient)
+		router, err := SetupRouter(log, env, opa, oas, sdk, mongoClient)
 		require.NoError(t, err, "unexpected error")
 		t.Run("/-/rbac-ready", func(t *testing.T) {
 			w := httptest.NewRecorder()
